Move generate-arg-types into standalone package
