+++ .gitignore
@@ -36,6 +36,9 @@ yarn-error.log*
 # turbo
 .turbo
 
+# cli utils
+.bin
+
 # data
 /data
 *.db
