+++ packages/generate-arg-types/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  // Extends can't use node's resolution rn https://github.com/dividab/tsconfig-paths/issues/209#issuecomment-1177261091
+  "extends": "../../node_modules/@webstudio-is/tsconfig/base.json",
+  "include": ["src"],
+  "compilerOptions": {
+    "module": "ES2020",
+    "target": "ES2020",
+    "strict": true,
+    "baseUrl": ".",
+    "rootDir": "./src",
+    "outDir": "./.bin"
+  },
+  // this is auto-generated by update-ts-project-references.ts
+  "references": []
+}
