+++ packages/react-sdk/package.json
@@ -8,7 +8,7 @@
   "scripts": {
     "dev": "tsup --watch",
     "build": "yarn build:lib",
-    "build:args": "tsx ./bin/generate-arg-types.ts './src/components/*.tsx !./src/**/*.stories.tsx !./src/**/*.ws.tsx'",
+    "build:args": "generate-arg-types './src/components/*.tsx !./src/**/*.stories.tsx !./src/**/*.ws.tsx' && prettier --write \"**/*.props.json\"",
     "typecheck": "tsc --noEmit",
     "build:cjs": "tsc --module commonjs --outDir lib/cjs && find lib/cjs -name '*.js' | while read NAME; do mv $NAME ${NAME%.js}.cjs; done",
     "build:lib": "rm -fr lib tsconfig.tsbuildinfo && tsc && yarn build:cjs",
@@ -33,15 +33,14 @@
     "@storybook/manager-webpack4": "^6.5.6",
     "@storybook/react": "^6.5.6",
     "@storybook/testing-library": "^0.0.11",
-    "@types/fs-extra": "^9.0.13",
     "@types/node": "^17.0.21",
     "@webstudio-is/jest-config": "*",
     "@webstudio-is/prisma-client": "*",
+    "@webstudio-is/generate-arg-types": "*",
     "babel-loader": "^8.2.5",
     "esbuild": "^0.14.25",
     "esbuild-node-externals": "^1.4.1",
     "react": "^17.0.2",
-    "react-docgen-typescript": "^2.2.2",
     "react-dom": "^17.0.2",
     "remix-utils": "^4.1.0",
     "tsup": "^6.1.3",
